Move a modify_indexes::Db::remove_index into a db actor
This is a part of #3.

The change moves remove_index method from modify_indexes into a db
actor. After that removal modify_indexes actor will be empty, so it is
removed. Next paches will move other methods into a db index actor.
